Распределит actions в тематические директории
