+++ project/src/index.tsx
@@ -1,6 +1,6 @@
 import App from './components/app/app';
 import ErrorMessage from './components/error-message/error-message';
-import { fetchOfferAction } from './store/api-actions';
+import { fetchOfferAction } from './store/offers/api-actions';
 import { Provider } from 'react-redux';
 import React from 'react';
 import ReactDOM from 'react-dom';
