+++ project/src/pages/offers-sort/offers-sort.tsx
@@ -1,7 +1,7 @@
 import { OffersType } from '../../types/offers';
 import { offersSortTypes, OffersSortType } from '../../const';
 import { OffersSortTypeKey } from '../../types/offers-sort';
-import { setOffers, setOffersSortType } from '../../store/action';
+import { setOffers, setOffersSortType } from '../../store/offers/action';
 import { sortOffersByType } from '../../utils/sort';
 import { useState } from 'react';
 import { useAppDispatch, useAppSelector } from '../../hooks';
