+++ project/src/store/offer-item/action.ts
@@ -1,0 +1,9 @@
+import { createAction } from '@reduxjs/toolkit';
+import { CommentsType } from '../../types/comments';
+import { OfferType, OffersType } from '../../types/offers';
+
+export const loadOfferItem = createAction<OfferType>('data/loadOfferItem');
+
+export const loadOffersNearby = createAction<OffersType>('data/loadOffersNearby');
+
+export const loadComments = createAction<CommentsType>('data/loadComments');
