+++ project/src/store/user/action.ts
@@ -1,0 +1,8 @@
+import { AuthorizationStatus, AppRoute } from '../../const';
+import { createAction } from '@reduxjs/toolkit';
+
+export const requireAuthorization = createAction<AuthorizationStatus>('user/requireAuthorization');
+
+export const redirectToRoute = createAction<AppRoute>('user/redirectToRoute');
+
+export const setError = createAction<string>('game/setError');
