+++ project/src/store/user/api-actions.ts
@@ -1,0 +1,45 @@
+import { AuthData } from '../../types/auth';
+import { createAsyncThunk } from '@reduxjs/toolkit';
+import { errorHandle } from '../../services/error-handle';
+import { store, api } from '..';
+import { saveToken, dropToken } from '../../services/token';
+import { TIMEOUT_SHOW_ERROR, APIRoute, AuthorizationStatus } from '../../const';
+import { UserData } from '../../types/user-data';
+import { setError, requireAuthorization } from './action';
+
+export const clearErrorAction = createAsyncThunk('game/clearError', () => {
+  setTimeout(() => store.dispatch(setError('')), TIMEOUT_SHOW_ERROR);
+});
+
+export const checkAuthAction = createAsyncThunk('user/checkAuth', async () => {
+  try {
+    await api.get(APIRoute.Login);
+    store.dispatch(requireAuthorization(AuthorizationStatus.Auth));
+  } catch (error) {
+    errorHandle(error);
+    store.dispatch(requireAuthorization(AuthorizationStatus.NoAuth));
+  }
+});
+
+export const loginAction = createAsyncThunk('user/login', async ({ login: email, password }: AuthData) => {
+  try {
+    const {
+      data: { token },
+    } = await api.post<UserData>(APIRoute.Login, { email, password });
+    saveToken(token);
+    store.dispatch(requireAuthorization(AuthorizationStatus.Auth));
+  } catch (error) {
+    errorHandle(error);
+    store.dispatch(requireAuthorization(AuthorizationStatus.NoAuth));
+  }
+});
+
+export const logoutAction = createAsyncThunk('user/logout', async () => {
+  try {
+    await api.delete(APIRoute.Logout);
+    dropToken();
+    store.dispatch(requireAuthorization(AuthorizationStatus.NoAuth));
+  } catch (error) {
+    errorHandle(error);
+  }
+});
